Finalización de crud de usuarios Inicio de crud de tickets

+++ proyectofinal/destroyuser.php
@@ -1,11 +1,16 @@
 <?php
+    session_start();
+    if (!isset($_SESSION['logged']) || $_SESSION['logged'] != true) {
+        header('Location: login.php');
+        return;
+    }
     require_once "core/class/User.php";
     
     $id = $_POST['id'];
     $user = new User;
     if($user->delete($id)){
-        //mensaje
+        $_SESSION['success'] = "Usuario eliminado de forma exitosa";
     }else{
-        //mensaje
+        $_SESSION['error'] = "Error, intente nuevamente y si persiste, comuniquese con el WebMaster";
     }
     header('Location: users.php');
